+++ Projeto-OO/src/app/Formulas_Calculo_Renda.java
@@ -2,7 +2,7 @@ package app;
 
 public interface Formulas_Calculo_Renda {
 
-	public abstract float calculoRegraIgualitaria();
+	public abstract double calculoRegraIgualitaria();
 	
-	public abstract float calculoRegraProporcional();
+	public abstract double calculoRegraProporcional();
 }
